converted connectivity to neotraits
git-svn-id: https://repo.thevirtualbrain.org/svn/tvb/trunk@8962 1c0e02f0-7929-43c0-8fb3-3293bf43b0d1

+++ scientific_library/tvb/basic/readers.py
@@ -200,7 +200,7 @@ class ZipReader(object):
         try:
             return self.read_array_from_file(file_name, dtype, skip_rows, use_cols, matlab_data_name)
         except ReaderException:
-            return numpy.array([])
+            return numpy.array([], dtype=dtype)
